+++ lib/index.ts
@@ -836,6 +836,30 @@ export default class Qsu {
 		return [...new Set(str)].join('');
 	}
 
+	static urlJoin(...args: string[]): string {
+		if (!args) {
+			return '';
+		}
+
+		const argLength = args.length;
+		let urlResult = '';
+
+		for (let i = 0; i < argLength; i += 1) {
+			if (
+				i === 0 ||
+				args[i].startsWith('/') ||
+				args[i].startsWith('?') ||
+				args[i].startsWith('&')
+			) {
+				urlResult += args[i];
+			} else {
+				urlResult += `/${args[i]}`;
+			}
+		}
+
+		return urlResult;
+	}
+
 	/*
 	 * Verify
 	 * */
@@ -1096,6 +1120,7 @@ export const {
 	strToNumberHash,
 	strUnique,
 	strToAscii,
+	urlJoin,
 	isObject,
 	isEqual,
 	isEqualStrict,