+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.3.3 (2024--)
 
 - `objFindItemRecursiveByKey`: Add `objFindItemRecursiveByKey` method
+- `urlJoin`: Add `urlJoin` method
 
 ## 1.3.2 (2023-12-28)